[REFACTOR:불필요한 메서드,변수삭제/클래스 디렉토리 변경(같은기능을하는 클래스끼리 같은디렉토리에 위치하게 변경)]
