+++ src/Data/StudentData.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 public class StudentData {
     protected int studentId;
     protected String studentName;
-
     // 1=Green, 2=Yellow, 3=Red
     int studentState;
     public StudentData(int studentId, String name, int state) {
@@ -13,18 +12,14 @@ public class StudentData {
         this.studentName = name;
         this.studentState = state;
     }
-
     ArrayList<Integer> requireAndChoice=new ArrayList<>();
     //학생이 무슨과목을듣는지 넣어놓는 리스트
     // subjectType : 0 = 필수, 1 = 선택
     // 등급 : A = 1, B = 2, C = 3, D = 4, F = 5, N = 6
-
-
     final int detailsSize=4;
     //배열크기가 바뀌면 detailsSize의 값을바꾸자.
     ArrayList<int[]> subjectList = new ArrayList<>();
     int[] subjectDetails = new int[detailsSize];
-
     public void subjectDetailsInput(int subjectId,int subjectType, int score,int round){
         subjectDetails = new int[detailsSize];
         //초기화 안해주면 고장남!!
@@ -50,9 +45,6 @@ public class StudentData {
         //inquiry클래스에서 사용할 SubjectList를 get하는 메서드추가 --multiverse22
         return subjectList;
     }
-    public void setSubjectList(int i,int newScore){
-        this.subjectList.get(i)[2]=newScore;
-    }
     public int getStudentState(){
         return studentState;
     }
@@ -61,9 +53,6 @@ public class StudentData {
     }
     public int getStudentId() {
         return this.studentId;
-    }
-    public void setStudentId(int studentName) {
-        this.studentId = studentName;
     }
     public String getStudentName() {
         return studentName;
